requirements.txt
# ================================================================
# Crypto Trading Strategy - Requirements
# ================================================================
# This file contains all required Python packages for the 
# crypto trading strategy system.

# Core Data Processing & Analysis
# ================================================================
numpy>=1.24.0
pandas>=2.0.0

# Machine Learning & Deep Learning
# ================================================================
scikit-learn>=1.3.0
tensorflow>=2.13.0

# Data Visualization
# ================================================================
matplotlib>=3.7.0
seaborn>=0.12.0

# Environment Management
# ================================================================
python-dotenv>=1.0.0

# HTTP Requests & API Communication
# ================================================================
requests>=2.28.0

# Trading Platform APIs
# ================================================================
alpaca-trade-api>=3.0.0

# Binance API (if using binance-python)
# Note: Using custom binance client, but these might be needed
# binance>=1.0.0
# python-binance>=1.0.17

# Development & Testing
# ================================================================
pytest>=7.4.0
jupyter>=1.0.0

# Data Processing Utilities
# ================================================================
pathlib

# Time & Date Handling  
# ================================================================
# datetime, time, timedelta - built into Python
pytz>=2023.3

# Type Hints & Annotations
# ================================================================
typing-extensions>=4.5.0

# Async Operations (for real-time data)
# ================================================================
asyncio
aiohttp>=3.8.0

# Performance & Monitoring
# ================================================================
psutil>=5.9.0
tqdm>=4.65.0

# Statistical Analysis
# ================================================================
scipy>=1.10.0
statsmodels>=0.14.0

# Technical Analysis
# ================================================================
ta>=0.10.2

# Configuration & Validation
# ================================================================
pydantic>=2.3.0

# Cryptographic Functions (for API security)
# ================================================================
cryptography>=41.0.0

# JSON Processing
# ================================================================
ujson>=5.8.0

# Caching & Persistence
# ================================================================
joblib>=1.3.0

# Error Handling & Monitoring
# ================================================================
sentry-sdk>=1.32.0

# Data Export/Import
# ================================================================
openpyxl>=3.1.0

# Web Scraping (optional for additional data sources)
# ================================================================
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database Support (for future data storage)
# ================================================================
sqlalchemy>=2.0.0

# Financial Data Sources (optional)
# ================================================================
yfinance>=0.2.0

# Logging Enhancement
# ================================================================
coloredlogs>=15.0

# Parallel Processing
# ================================================================
multiprocessing-logging>=0.3.4

# Advanced Visualization (optional)
# ================================================================
plotly>=5.15.0
bokeh>=3.2.0

# Performance Optimization (optional)
# ================================================================
numba>=0.57.0

# Time Series Analysis (advanced)
# ================================================================
arch>=6.2.0

# Timezone Handling
# ================================================================
tzlocal>=5.0

# Data Structures
# ================================================================
sortedcontainers>=2.4.0

# Regular Expressions (enhanced)
# ================================================================
regex>=2023.8.8

# Built-in Python modules (included for reference)
# ================================================================
# os, sys, json, logging, time, threading, signal, traceback
# datetime, timedelta, typing, dataclasses, enum, argparse
# decimal, pathlib, glob, re, warnings, urllib.parse
# hashlib, hmac, gzip, configparser, smtplib, sqlite3
